+++ lib/src/web_helper/web_save_excel_browser.dart
@@ -1,6 +1,7 @@
 import 'dart:js_interop';
 import 'dart:typed_data';
 
+import 'package:cross_file/cross_file.dart';
 import 'package:web/web.dart';
 
 // A wrapper to save the excel file in browser
@@ -25,5 +26,19 @@ class SavingHelper {
     anchor.remove();
     URL.revokeObjectURL(url);
     return val;
+  }
+
+  static XFile? generateXFile(List<int>? val, String fileName) {
+    if (val == null) return null;
+    final bytes = Uint8List.fromList(val);
+    return XFile.fromData(
+      bytes,
+      name: fileName,
+      length: bytes.length,
+      lastModified: DateTime.now(),
+      mimeType:
+          'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
+      path: fileName,
+    );
   }
 }